Refactor/250 : 프로필 카테고리별 정답률 집계 안되는 문제 해결, Quiz Json 업로드 Dto 수정 (#257)
* feat :  의존성 추가

* chore : 중복 api 제거로 불필요해진 코드 일부 제거

* chore : 퀴즈 Json 파일 업로드 시, 불필요한 format 제거

* refactor : 정답률 불러오기 해결, 문제 생성 DTO 변결

* chore : 의존성 버전 통일
