+++ cs25-batch/build.gradle
@@ -32,6 +32,7 @@ dependencies {
     implementation platform("software.amazon.awssdk:bom:2.25.39")
     implementation 'software.amazon.awssdk:sesv2'
     implementation 'software.amazon.awssdk:netty-nio-client'
+    implementation 'software.amazon.awssdk:auth'
 
     //Bucket4J
     implementation 'com.bucket4j:bucket4j_jdk17-core:8.14.0'
