--- cs25-entity/src/main/java/com/example/cs25entity/domain/quiz/dto/QuizSearchDto.java
@@ -1,17 +1,0 @@
-package com.example.cs25entity.domain.quiz.dto;
-
-import com.example.cs25entity.domain.quiz.enums.QuizLevel;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class QuizSearchDto {
-    private Long categoryId;
-    private QuizLevel level;
-
-    @Builder
-    public QuizSearchDto(Long categoryId, QuizLevel level) {
-        this.categoryId = categoryId;
-        this.level = level;
-    }
-}
