+++ cs25-entity/src/main/java/com/example/cs25entity/domain/quiz/repository/QuizCustomRepository.java
@@ -1,13 +1,10 @@
 package com.example.cs25entity.domain.quiz.repository;
 
-import com.example.cs25entity.domain.quiz.dto.QuizSearchDto;
 import com.example.cs25entity.domain.quiz.entity.Quiz;
 import com.example.cs25entity.domain.quiz.enums.QuizFormatType;
 import com.example.cs25entity.domain.quiz.enums.QuizLevel;
 import java.util.List;
 import java.util.Set;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
 public interface QuizCustomRepository {
 
@@ -16,5 +13,4 @@ public interface QuizCustomRepository {
         Set<Long> solvedQuizIds,
         List<QuizFormatType> targetTypes);
 
-    Page<Quiz> searchQuizzes(QuizSearchDto condition, Pageable pageable);
 }
