+++ cs25-service/build.gradle
@@ -36,8 +36,7 @@ dependencies {
     implementation platform("software.amazon.awssdk:bom:2.25.39")
     implementation 'software.amazon.awssdk:sesv2'
     implementation 'software.amazon.awssdk:netty-nio-client'
-    implementation 'io.github.resilience4j:resilience4j-ratelimiter:2.1.0'
-
+    implementation 'software.amazon.awssdk:auth'
     // Jwt
     implementation 'io.jsonwebtoken:jjwt-api:0.12.6' //service
     implementation 'io.jsonwebtoken:jjwt-impl:0.12.6' //service
