+++ cs25-service/src/main/java/com/example/cs25service/domain/admin/dto/request/CreateQuizDto.java
@@ -1,7 +1,5 @@
-package com.example.cs25service.domain.quiz.dto;
+package com.example.cs25service.domain.admin.dto.request;
 
-import com.example.cs25entity.domain.quiz.entity.QuizCategory;
-import com.example.cs25entity.domain.quiz.enums.QuizLevel;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
@@ -11,9 +9,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class CreateQuizDto {
-
-    @NotBlank(message = "문제 타입은 필수입니다.")
-    private String type;
 
     @NotBlank(message = "문제는 필수입니다.")
     private String question;
@@ -32,9 +27,8 @@ public class CreateQuizDto {
     private String level;
 
     @Builder
-    public CreateQuizDto(String type, String question, String choice, String answer, String commentary,
+    public CreateQuizDto(String question, String choice, String answer, String commentary,
         String category, String level) {
-        this.type = type;
         this.question = question;
         this.choice = choice;
         this.answer = answer;
