+++ cs25-service/src/main/java/com/example/cs25service/domain/admin/service/QuizAdminService.java
@@ -12,7 +12,7 @@ import com.example.cs25entity.domain.userQuizAnswer.repository.UserQuizAnswerRep
 import com.example.cs25service.domain.admin.dto.request.QuizCreateRequestDto;
 import com.example.cs25service.domain.admin.dto.request.QuizUpdateRequestDto;
 import com.example.cs25service.domain.admin.dto.response.QuizDetailDto;
-import com.example.cs25service.domain.quiz.dto.CreateQuizDto;
+import com.example.cs25service.domain.admin.dto.request.CreateQuizDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
